Adds npmignore file.
The lib directory doesn't need to be stored on github as it can be
regenerated from src. However, adding it to the .gitignore file meant it
wouldn't get published to npm. Adding a .npmignore file allows us to
exclude the lib directory from  github, but continue to publish it to
npm.