Adds viewer position reset if position array empty
For some reason the neuroglancer window does not reset the position of
the camera if the position is set to an empty array. This update checks
the position value and if it is found to be an empty array it calls the
viewer.position.reset() function which resets the view to the center of
the volume.